fix old bsp won't decompile
Added quakelive support in q3map2 2.5.17 introduced a new bsp lump,
that contains advertisements data, but didn't bother to increase
protocol version, or rather it is increased relatively to quake 3,
but still has same protocol number as rtcw and et. Whilst trying to
decompile olderbsps, that have used older map compiler, will lead
to a "funny lump size" error, because advertisement lump being
processed despite the fact there is no more data left. This doesn't
happen on a maps compiled with newer compiler versions due to the
fact it still writes an empty lump information to the end of bsp,
so the compiler always knows there is nothing left to fetch.

This commit adds a bsp file size check to make sure that after prelast
lump, there is still a data left to process, otherwise it doen't bother
to continue.